recon: rewrite 11 red functions from disassembly -- 1 red->green, ~2350 bytes closer
The RED (low-match) functions were reconstruction gaps, not compiler walls: our C had
wrong or missing logic, which is source-addressable. A fleet of agents rebuilt each from
the target disassembly, keeping only semantics-preserving reconstructions.

  find_rec_by_field8      27% -> MATCHED (reloc-aware)   volatile cursor globals + operand order
  agent_hud_render        49% -> 77% (1224->546)         unsigned-short counter stops strength-reduction
  mission_target_resolve  27% -> 55% (2692->1669)        rebuilt the switch(g_e120) action-dispatch tail
  store_squad_loadout     51% -> 70% (702->439)          record-first indexing restores REC-reuse
  new_campaign_reset      46% -> 58% (810->636)          duplicated zero-loops into both fund branches
  parse_hex_arg           16% -> code matched            hex jump-table + strlen intrinsic rebuilt
  path_probe_y            63% -> 67% (121->108)          tile-byte hoist matches column lookup
  FUN_0002d5b8            63% -> 67% (120->107)          same (map passability twin)
  economy_daily_tick      65% -> 66% (336->322)          reload-schedule fix
  draw_filled_shape       56% -> 58% (732->697)          int-promote homes py in EDI
  text_width_kern         near-miss improved             short-typing flipped the register role

The pattern: functions red because logic was MISSING (ours << target size) pay off big;
functions already built but wall-bound give only a few bytes. Each header rewritten to
document what the function actually does. center_string_16 reclassified (was a stale amber
tag on an already-exact function). Treemap 356 -> 357 matched, 62.5% bytes.

+++ src/ui/agent_hud_render.c
@@ -40,10 +40,27 @@
  * field, cache offset, y position and colour pair; they are captured here by the
  * IPA() macro (expands inline, exactly as the compiler emitted them -- no call).
  *
- * STATUS: PARKED. Register-role / accumulator walls (the repeated `*55 IDIV 255`
- * chains and the two-segment redraw's stack-shuffled colour push) are not
- * source-reachable, and the wrong manifest size (2080 vs real 2387) makes
- * match_reloc length-mismatch regardless. Recipe: -4s -oneatx -zp8 -s -zq.
+ * STATUS: 77% match, EDIT-DIST 546 (was 49% / 1224). The big win was typing the
+ * loop counter `i` as `unsigned short` (was `int`): the 16-bit counter makes every
+ * `i*0xb`/`i*0x12` index recompute from `di` (mov ax,di / imul) instead of the
+ * compiler strength-reducing them into running induction pointers (esi+=0xb,
+ * ebx+=0x12). That single change realigned the whole first loop -- register roles
+ * (esi=p, edi=i, ebx=mid accumulator, ecx=i*0xb), the 16-bit `cmp di,4 / jb` loop
+ * tests, and the blink table's `cmp cx,8 / jb`. The blink loop and the entire
+ * second (aux progress-bar) pass are now byte-identical.
+ *
+ * REMAINING GAP (~546) is a Watcom-9.5 register/stack-slot tie, not source-reachable:
+ *   - Frame is 0xc vs target's 8: Watcom reserves an unused home slot for `mid`
+ *     (which actually lives in ebx), shifting every IPA/health stack temp from
+ *     offsets {0,4} to {4,8} -- ~1 byte per esp-disp across three bars + health.
+ *   - Two-segment redraw: target reloads lo/hi from their slots and recomputes
+ *     `v-mid` in the shared tail; ours CSEs `lo-mid` and spills/reuses it. Forcing
+ *     recompute (shared v/c tail, or `volatile` on lo/hi) regresses globally.
+ *   - Muzzle-flash: `st` lands in dl vs target's bl, flipping the two flag stores'
+ *     order and the const register (bh vs dl).
+ * All three resisted decl-order, health-var-reuse, volatile, and shared-tail edits.
+ * The wrong manifest size (2080 vs real 2387) still forces a length-mismatch in
+ * match_reloc regardless. Recipe: -4s -oneatx -zp8 -s -zq.
  */
 
 extern unsigned char  g_blink_tick;      /* blink tick counter */
@@ -98,7 +115,7 @@ extern void draw_slot_record_chain(int p, unsigned short idx, int x, int y);
 
 void agent_hud_render(void)
 {
-    int i;
+    unsigned short i;
     int lo, mid, hi;
     unsigned char *p;
     unsigned short id;
